refactor: 디렉토리 구조를 step1 방식으로 되돌린다(pages 디렉토리 삭제)
